+++ Changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [UNRELEASED]
+### Added
+ - Time component.
 
 ### Added
  - Allow passing in a baseUrl for forms on a different path.